test: move GH-29558 cases into crypto.key-objects.test.ts
Not a regression (never worked in any Bun release — BoringSSL has always
rejected this encoding), so test/regression/issue/ is the wrong home per
the repo's test-layout convention. Fold into the existing KeyObject tests.
